refactor: extract to useAutoStatusUpdate hook
